requirements.txt
setuptools==59.5.0
numpy==1.22
matplotlib==3.5
scipy==1.8
scikit_learn==1.1
torch==1.11
torchvision
torchaudio
pytorch_lightning==1.6
omegaconf
hydra-core==1.2
librosa==0.9
tqdm
einops
tensorflow
tensorboard
dill
termcolor
GPUtil
crepe
av==9.2
git+https://github.com/caillonantoine/cached_conv.git#egg=cached_conv
git+https://github.com/domkirke/acids_transforms.git
git+https://github.com/domkirke/lardon.git
git+https://github.com/domkirke/trajectories.git@master
pytest